requirements.txt
bleach==3.1.5
Django==3.1
easy-thumbnails==2.7
django-extensions==3.0.5
django-tags-input==5.0.0
djangorestframework==3.11.1
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-environ==0.4.5
mysqlclient==1.4.2.post1
django-map-widgets==0.2.2
django-model-utils==4.0.0
Pillow==5.1.0
social-auth-app-django==3.4.0
django-fullurl==1.0
user-agents==2.1
django-user-agents==0.4.0
IP2Location==8.4.1
ipaddress==1.0.23
beautifulsoup4==4.9.1
django-ckeditor==6.0.0
django-import-export==2.2.0
num2words==0.5.10
django-su==0.9.0
django-cookie-law==2.0.3
django-post-office==3.4.1
django-ses==1.0.3
requests==2.22.0
django-htmlmin==0.11.0
